% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_helpers.R
\name{api_headers}
\alias{api_headers}
\title{Construct headers for GitHub API requests}
\usage{
api_headers()
}
\value{
Named vector of HTTP headers
}
\description{
Construct headers for GitHub API requests
}
\keyword{internal}
