% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_chrr_counties.R
\name{list_chrr_counties}
\alias{list_chrr_counties}
\title{List available counties for one or more states}
\usage{
list_chrr_counties(state)
}
\arguments{
\item{state}{Character vector. State abbreviation(s) (e.g. "WI") or full state name(s) (e.g. "Wisconsin").}
}
\value{
A tibble with columns: state, county, statecode, countycode, fipscode.
}
\description{
List available counties for one or more states
}
\examples{
\dontrun{
list_chrr_counties("WI")
list_chrr_counties(c("WI", "MN"))
list_chrr_counties(c("Wisconsin", "Minnesota"))
list_chrr_counties(c("wi", "mn"))
}
}
