% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_chrr_years.R
\name{list_chrr_years}
\alias{list_chrr_years}
\title{List available CHRR data years}
\usage{
list_chrr_years()
}
\value{
A vector of available years (numeric).
}
\description{
Uses the GitHub API to find available years in the CHRR measure calcs repo.
}
\examples{
\dontrun{
list_chrr_years()
}
}
