% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_chrr_measures.R
\name{list_chrr_measures}
\alias{list_chrr_measures}
\title{List available CHR&R measures for a given release year}
\usage{
list_chrr_measures(release_year = 2023)
}
\arguments{
\item{release_year}{Numeric year. This corresponds to the year of CHR&R's annual data release.}
}
\value{
A tibble with measure_id and measure_name.
}
\description{
Downloads metadata for measures and returns their IDs and names.
}
\examples{
\dontrun{
list_chrr_measures(2023)
}
}
